+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased changes
 
 - Fix a bug in computing the number of missed rounds in the event of a timeout.
+- Fix a bug where the suspended status of a baker pool would be omitted when it was suspended.
 
 ## 8.0.0